добавил closeSessionFactory, поправил обработку исключений в соответствии с замечаниями, поправил инкапсуляцию в UserServiceImpl, добавил вывод в консоль в методе getAllUsers.

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -1,6 +1,7 @@
 package jm.task.core.jdbc.dao;
 
 import jm.task.core.jdbc.model.User;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
@@ -10,10 +11,10 @@ import java.util.List;
 import static jm.task.core.jdbc.util.Util.getSessionFactory;
 
 public class UserDaoHibernateImpl implements UserDao {
+
     public UserDaoHibernateImpl() {
 
     }
-
 
     @Override
     public void createUsersTable() {
@@ -31,80 +32,91 @@ public class UserDaoHibernateImpl implements UserDao {
             query.executeUpdate();
 
             transaction.commit();
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             throw new RuntimeException("Ошибка создания таблицы", e);
-
         }
     }
 
     @Override
     public void dropUsersTable() {
-
         try (Session session = getSessionFactory().openSession()) {
             Transaction transaction = session.beginTransaction();
+
             String sql = "DROP TABLE IF EXISTS users";
             Query query = session.createSQLQuery(sql).addEntity(User.class);
             query.executeUpdate();
+
             transaction.commit();
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             throw new RuntimeException("Ошибка удаления таблицы", e);
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-
+        Transaction transaction = null;
         try (Session session = getSessionFactory().openSession()) {
-            Transaction transaction = session.beginTransaction();
+            transaction = session.beginTransaction();
+
             User user = new User(name, lastName, age);
             session.save(user);
-            transaction.commit();
 
-        } catch (Exception e) {
+            transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
             throw new RuntimeException("Ошибка добавления пользователя", e);
         }
-
     }
 
     @Override
     public void removeUserById(long id) {
+        Transaction transaction = null;
         try (Session session = getSessionFactory().openSession()) {
-            Transaction transaction = session.beginTransaction();
-            User user = session.get(User.class, id);
-            session.delete(user);
-            transaction.commit();
+            transaction = session.beginTransaction();
 
-        } catch (Exception e) {
+            User user = session.get(User.class, id);
+            if (user != null) {
+                session.delete(user);
+                transaction.commit();
+            } else {
+                transaction.rollback();
+                throw new RuntimeException("Пользователь с id " + id + " не найден");
+            }
+        } catch (HibernateException e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
             throw new RuntimeException("Ошибка удаления пользователя", e);
         }
-
     }
-
 
     @Override
     public List<User> getAllUsers() {
         try (Session session = getSessionFactory().openSession()) {
-            Transaction transaction = null;
-            try {
-                transaction = session.beginTransaction();
-                List<User> users = session.createQuery("from User").getResultList();
-                return users;
-            } catch (Exception e) {
-                if (transaction != null) {
-                    transaction.rollback();
-                }
-                throw e;
-            }
+            List <User> users = session.createQuery("from User", User.class).getResultList();
+            System.out.println(users);
+            return users;
+        } catch (HibernateException e) {
+            throw new RuntimeException("Ошибка получения списка пользователей", e);
         }
     }
 
     @Override
     public void cleanUsersTable() {
+        Transaction transaction = null;
         try (Session session = getSessionFactory().openSession()) {
-            Transaction transaction = session.beginTransaction();
-            session.createQuery("delete from User").executeUpdate();
-            transaction.commit();
-        }
+            transaction = session.beginTransaction();
 
+            session.createQuery("delete from User").executeUpdate();
+
+            transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
+            throw new RuntimeException("Ошибка очистки таблицы пользователей", e);
+        }
     }
 }
